+++ src/fsfw/cfdp/CfdpHandler.h
@@ -28,7 +28,7 @@ class CfdpHandler : public ExecutableObjectIF,
   virtual ReturnValue_t handleRequest(store_address_t storeId);
 
   virtual ReturnValue_t initialize() override;
-  virtual uint16_t getIdentifier() override;
+  virtual uint32_t getIdentifier() override;
   MessageQueueId_t getRequestQueue() override;
   ReturnValue_t performOperation(uint8_t opCode) override;
